WIP - Make default serializer configurable
This commit loads a new configuration key allowing the user to
set the default content-type and serializer.

    # config.yml
    serializer: Mutable
    engines:
        serializer:
            Mutable:
                default:
                    'application/custom' : Custom

If set, Dancer2::Serializer::Mutable overrides the content_type
attribute using the key, else sets the previous default,
application/json.

During serialization and deserialization, if a match isn't found in
the mapping, the new key and value will be used if set, else will
default to the previous values, application/json (content_type)
and JSON (serializer).